requirements.txt
aiofiles==25.1.0
albumentations==0.4.5
altair==5.5.0
aniso8601==10.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
attrs==25.4.0
blinker==1.9.0
boto3==1.40.54
botocore==1.40.54
cachetools==6.2.1
certifi==2022.12.7
cffi==2.0.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
Cython==3.1.4
dynaconf==3.2.12
effdet==0.3.0
efficientnet_pytorch==0.7.1
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grad-cam==1.5.5
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hf-xet==1.1.10
hpack==3.0.0
hstspreload==2025.1.1
httptools==0.7.1
httpx==0.9.3
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imgaug==0.2.6
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==4.7.6
narwhals==2.9.0
networkx==3.2.1
Nuitka==2.8.4
numpy==1.26.4
omegaconf==2.3.0
onnx==1.19.1
onnxruntime-gpu==1.19.2
onnxsim==0.4.36
onnxslim==0.1.73
opencv-python==4.8.1.78
opencv-python-headless==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.35.1
polars-runtime-32==1.35.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.9.1
pydicom==2.1.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.27.1
s3transfer==0.14.0
safetensors==0.6.2
sanic==19.12.2
sanic-restful-api==0.0.8
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
statsd==4.0.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.6.13
toml==0.10.2
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.5.2
tqdm==4.67.1
ttach==0.0.3
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3==1.25.11
uvicorn==0.24.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
zipp==3.23.0
zstandard==0.25.0
